[level 2] Title: 멀리 뛰기, Time: 0.16 ms, Memory: 83.8 MB -BaekjoonHub

+++ 프로그래머스/2/12914. 멀리 뛰기/멀리 뛰기.java
@@ -1,6 +1,6 @@
 class Solution {
     public long solution(int n) {
-        long[] dp = new long[n + 1];
+        int[] dp = new int[n + 1];
         dp[0] = 1;
         dp[1] = 1;
